+++ api_tests/server_api_test.py
@@ -1,4 +1,5 @@
 import json
+import re
 
 from conftest import requires_env, auth_headers
 
@@ -9,7 +10,7 @@ def test__get_info(base_url, session):
 
     assert resp.status_code == 200
     data = json.loads(resp.content)
-    assert data['version'] >= '1.0.0'
+    assert data['version'] >= '1.3.0'
     assert int(data['bucket_count']) > 0
     assert int(data['uptime']) >= 0
     assert int(data['latest_record']) >= 0
@@ -17,7 +18,7 @@ def test__get_info(base_url, session):
 
     assert data['defaults']['bucket'] == {'max_block_records': '1024', 'max_block_size': '64000000', 'quota_size': '0',
                                           'quota_type': 'NONE'}
-    assert resp.headers['server'] == "ReductStorage"
+    assert re.match(r"ReductStore 1\.\d+\.\d+", resp.headers['server'])
     assert resp.headers['Content-Type'] == "application/json"